Add list/help commands, fix DB connect and logging
Add printAll and help/list handling in main and expand messaging map with help, helpPrompt and noEntries. Fix DB connection shadowing by assigning sql.Open result to the package-level db variable. Adjust selectAll/slct row iteration and include id in scans to properly read results. Change dateToStr to accept uint and convert to int64 for time.Unix. Reduce runtime.Caller depth in MarkFunc from 2 to 1. Also set running=false in stop() and update a prompt string (toStop). These changes tidy up runtime logging, DB usage, and user-facing commands/messages.

+++ conversions.go
@@ -23,10 +23,10 @@ func secToStr(sec uint) string {
 	)
 }
 
-func dateToStr(date int64) string {
+func dateToStr(date uint) string {
 	defer d.MarkFunc()
 
-	dateTime := time.Unix(date, 0)
+	dateTime := time.Unix(int64(date), 0)
 	return fmt.Sprintf("%d/%d/%d", dateTime.Day(), dateTime.Month(), dateTime.Year())
 }
